requirements.txt
absl-py==2.1.0
filelock==3.15.4
fsspec==2024.6.1
grpcio==1.64.1
Jinja2==3.1.4
Markdown==3.6
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.0.0
pandas==2.2.2
pillow==10.4.0
protobuf==4.25.3
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
sympy==1.13.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
torch==2.3.1
torchsummary==1.5.1
torchvision==0.18.1
typing_extensions==4.12.2
tzdata==2024.1
Werkzeug==3.0.3